Fix an issue for when new searches happen when index is not 0
Before, when a search has already been sent back, the user is currently
not viewing the first list of results, and a new search is then sent,
it would go to that page of that request instead of the first set of
results.

This issue has been fixed by reverting the index to 0 when new requests
are formed.
Index has been merged together with search to make it an object.